geometry restructuring
Restructured geometry a bit, made the distinction between
points and vectors clearer, switched a bunch of stuff to
float32, and made Neighbors for a square grid more like
the old modus's "Splash" on hex grids, which will get
implemented next for hex grids.

+++ g/context.go
@@ -61,7 +61,7 @@ func (c *Context) DrawSize() (int, int) {
 	}
 }
 
-func (c *Context) Render(screen *ebiten.Image, fn func(*ebiten.Image, float64)) {
+func (c *Context) Render(screen *ebiten.Image, fn func(*ebiten.Image, float32)) {
 	if c.multisample {
 		c.fsaa.Fill(color.RGBA{0, 0, 0, 0})
 		fn(c.fsaa, 2)
